♻️ refactor : ExpiredStudies 데이터타입 디렉토리 변경 -> study-service-dto 내부로 변경 (#97)
