+++ src/main/java/com/devcourse/checkmoi/domain/study/dto/StudyResponse.java
@@ -1,6 +1,5 @@
 package com.devcourse.checkmoi.domain.study.dto;
 
-import com.devcourse.checkmoi.domain.study.dto.StudyResponse.ExpiredStudies;
 import com.devcourse.checkmoi.domain.study.dto.StudyResponse.MyStudies;
 import com.devcourse.checkmoi.domain.study.dto.StudyResponse.Studies;
 import com.devcourse.checkmoi.domain.study.dto.StudyResponse.StudyAppliers;
@@ -20,8 +19,7 @@ import lombok.Builder;
 public sealed interface StudyResponse permits
     StudyInfo, StudyDetail, StudyDetailWithMembers,
     Studies, StudyBookInfo, StudyUserInfo,
-    StudyAppliers, MyStudies,
-    ExpiredStudies {
+    StudyAppliers, MyStudies {
 
     record StudyInfo(
         Long id,
@@ -131,11 +129,5 @@ public sealed interface StudyResponse permits
         public MyStudies {
 
         }
-    }
-
-    record ExpiredStudies(
-        List<Long> studies
-    ) implements StudyResponse {
-
     }
 }
