+++ src/main/java/com/devcourse/checkmoi/domain/study/service/dto/ExpiredStudies.java
@@ -1,0 +1,9 @@
+package com.devcourse.checkmoi.domain.study.service.dto;
+
+import java.util.List;
+
+public record ExpiredStudies(
+    List<Long> studies
+) {
+
+}
