+++ src/main/java/com/devcourse/checkmoi/global/scheduler/ScheduleManager.java
@@ -1,8 +1,8 @@
 package com.devcourse.checkmoi.global.scheduler;
 
-import com.devcourse.checkmoi.domain.study.dto.StudyResponse.ExpiredStudies;
 import com.devcourse.checkmoi.domain.study.model.StudyMemberStatus;
 import com.devcourse.checkmoi.domain.study.model.StudyStatus;
+import com.devcourse.checkmoi.domain.study.service.dto.ExpiredStudies;
 
 public interface ScheduleManager {
